option for auth of only specified plugin routes
